backend/requirements.txt
fastapi==0.111.0
pydantic-settings==2.2.1
uvicorn[standard]==0.29.0
sqlalchemy==2.0.30
alembic==1.13.1
PyMySQL==1.1.1
cryptography==42.0.5
python-jose[cryptography]==3.3.0
bcrypt
python-dotenv==1.0.1
python-multipart==0.0.9
qrcode[pil]==7.4.2
reportlab==4.1.0
Pillow==10.3.0
aiofiles==23.2.1
